requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow
torch
timm
torchvision
python-multipart
